split 3 upload steps to own file, rename TryUpload to Upload()
